fix: Take unroller timesteps as size_t
The unroller already stored times as size_t in var_times_ and returned
them as size_t from get_var_time and get_curr_time, but took them as
unsigned int, so callers holding a container size narrowed it on the way
in and widened it back on the way out.

+++ python/pono/pono.pyx
@@ -437,7 +437,7 @@ cdef class Unroller:
     def __dealloc__(self):
         del self.cu
 
-    def at_time(self, Term t, unsigned int k):
+    def at_time(self, Term t, size_t k):
         cdef Term term = Term(self._solver)
         term.ct = dref(self.cu).at_time(t.ct, k)
         return term
